Partly done full OAuth 2 integration (still need to fix it for client aim service)

+++ food-server/src/main/java/com/bahdanau/food/service/FoodService.java
@@ -15,27 +15,39 @@ public class FoodService {
     private final ModelMapper modelMapper;
     private final FoodRepository foodRepository;
 
-    public List<Food> getAllFood(String userId) {
-        return foodRepository.findAllByUserId(userId);
+    public List<Food> getAllFood(String userEmail) {
+        return foodRepository.findAllByUserEmail(userEmail);
     }
 
-    public List<Food> getFoodByCategory(String foodCategory, String userId) {
-        return foodRepository.findAllByFoodCategoryAndUserId(foodCategory.toUpperCase(), userId);
+    public List<Food> getFoodByCategory(String foodCategory, String userEmail) {
+        return foodRepository.findAllByFoodCategoryAndUserEmail(foodCategory.toUpperCase(), userEmail);
     }
 
-    public List<Food> getAllByName(String name, String userId) {
-        return foodRepository.findAllByNameContainingAndUserId(name, userId);
+    public List<Food> getAllByName(String name, String userEmail) {
+        return foodRepository.findAllByNameContainingAndUserEmail(name, userEmail);
     }
 
-    public Food addFoodItem(FoodDto foodDto) {
-        return foodRepository.insert(modelMapper.map(foodDto, Food.class));
+    public Food addFoodItem(FoodDto foodDto, String userEmail) {
+        Food foodItem = modelMapper.map(foodDto, Food.class);
+        foodItem.setUserEmail(userEmail);
+        return foodRepository.insert(foodItem);
     }
 
-    public Food updateFoodItem(FoodDto foodDto) {
+    public Food updateFoodItem(FoodDto foodDto, String userEmail) throws IllegalAccessException {
+        boolean foodItemOwnedByAnotherUser = foodRepository.findById(foodDto.getId()).stream()
+                .noneMatch(food -> food.getUserEmail().equals(userEmail));
+        if(foodItemOwnedByAnotherUser) {
+            throw new IllegalAccessException("This food item is attached to another user!");
+        }
         return foodRepository.save(modelMapper.map(foodDto, Food.class));
     }
 
-    public void removeFoodItem(String id) {
+    public void removeFoodItem(String id, String userEmail) throws IllegalAccessException {
+        boolean foodItemOwnedByAnotherUser = foodRepository.findById(id).stream()
+                .noneMatch(food -> food.getUserEmail().equals(userEmail));
+        if(foodItemOwnedByAnotherUser) {
+            throw new IllegalAccessException("This food item is attached to another user!");
+        }
         foodRepository.deleteById(id);
     }
 }
